feat: Sicherung & Umzug, Deinstallation, Datenbank-Verkleinerung (v3.4.1 / Helper 3.1.35)
Sicherung & Umzug
- Wöchentliche Sicherung von Datenbank und Einstellungen an einen frei wählbaren
  Ort: VACUUM INTO (konsistent bei laufendem Server), quick_check vor dem
  Einwechseln, genau eine Kopie -- die bisherige bleibt, bis die neue vollständig
  geschrieben ist.
- Einspielen ersetzt Datenbank und Einstellungen, legt vorher eine
  Sicherheitskopie an und schreibt fehlende Pfade auf den neuen Ort um. An
  240'308 Dokumenten geprüft, inklusive Pfadumschreibung.
- Mail-Passwörter sind bewusst nicht Teil der Sicherung; fehlende Passwörter
  werden in den Einstellungen und in der Diagnose sichtbar markiert.
- Der Server startet sich nach dem Einspielen selbst neu, damit der neue
  Bestand aktiv wird.

Deinstallation
- "Archivio Deinstallieren.app" führt mit macOS-Dialogen durch das Entfernen,
  ohne Terminal. Zwei Betriebsarten: komplett entfernen oder zum Arbeitsplatz
  machen (Helper bleibt).
- Nichts wird endgültig gelöscht, alles landet im Papierkorb.

Datenbank-Verkleinerung (Migrationen 026-028)
- documents_fts ohne content-Spalte: die Kopie aller Dokumenttexte wurde nie
  durchsucht, die Volltextsuche läuft über chunks_fts.
- Embeddings als float16 statt float32; Trefferunterschiede betreffen
  nachweislich nur Gleichstände.
- created_at je Chunk entfernt (nirgends gelesen).
- An der echten Produktivdatenbank gemessen: 5,97 -> 4,46 GB (-25,3 %),
  Migration 2,9 min.

Zuverlässigkeit
- Einmal-Start-Sperre: der Server konnte über LaunchAgent UND ein altes
  Anmeldeobjekt doppelt starten, beide Instanzen schossen sich gegenseitig ab.
- Migrationen laufen im Hintergrund, mit Warteseite statt toter Verbindung.
  Vorher hätte der Watchdog eine Migration über 60 s mitten drin abgebrochen.
- Log-Rotation (5 MB x 3, INFO) statt unbegrenzt wachsender DEBUG-Logs.
- Der Helper sucht den Server neu, wenn die gespeicherte Adresse dauerhaft
  tot ist -- nötig nach einem Serverumzug.
- Notarisierung übersteht abgerissene Verbindungen statt den Build zu verwerfen.

Entwicklungsumgebung auf Python 3.14 mit derselben SQLite-Version wie das
ausgelieferte Bundle (vorher 3.9 mit SQLite 3.34).

446 Tests.

+++ scanner/embedder.py
@@ -116,7 +116,7 @@ def embed_document_chunks(conn: sqlite3.Connection, document_id: int) -> int:
             for row, vec in zip(batch, vecs):
                 conn.execute(
                     "UPDATE document_chunks SET embedding = ? WHERE id = ?",
-                    (vec.tobytes(), row["id"])
+                    (vec.astype(np.float16).tobytes(), row["id"])
                 )
         total += len(batch)
     return total
@@ -161,8 +161,17 @@ def vector_search(
     if not rows:
         return []
 
+    # Gespeichert wird float16 (halber Platz, siehe Migration 027) -- der Genauigkeits-
+    # verlust ist beim Kosinusvergleich L2-normierter Vektoren nicht messbar. Der
+    # Anfragevektor bleibt float32; numpy hebt die Matrix beim Skalarprodukt an, so
+    # dass auf der Anfrageseite nichts verlorengeht.
+    #
+    # WICHTIG: der Datentyp steht nicht in der Datenbank. np.array() unten baut aus
+    # allen Zeilen EINE Matrix -- eine einzige Zeile im falschen Format ergibt eine
+    # andere Vektorlänge und lässt die gesamte semantische Suche kippen. Migration 027
+    # stellt deshalb alle Zeilen in einem Zug um.
     embeddings = np.array(
-        [np.frombuffer(r["embedding"], dtype=np.float32) for r in rows]
+        [np.frombuffer(r["embedding"], dtype=np.float16) for r in rows]
     )
     scores   = embeddings @ query_vec          # cosine similarity (vecs already normalised)
     top_idx  = np.argsort(scores)[::-1][:limit]
